ボタン状態処理の関数化
