+++ apps/frontend/src/pages/login/hooks/useBannerClose.ts
@@ -1,0 +1,14 @@
+export const useBannerCloseHandler = (el: HTMLElement): void => {
+  const bannerCloseBtn = el.querySelector(
+    "#bannerCloseBtn",
+  ) as HTMLButtonElement | null;
+
+  if (!bannerCloseBtn) return;
+
+  bannerCloseBtn.addEventListener("click", () => {
+    const banner = el.querySelector("#loginBanner");
+    if (banner) {
+      banner.classList.add("hidden");
+    }
+  });
+};
